part4: 0-21 (except 6-7)

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,23 @@
 
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
+  return response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', async (request, response, next) => {
   const blog = new Blog(request.body)
 
   if (!blog.title && !blog.url) {
@@ -20,19 +30,42 @@ blogsRouter.post('/', async (request, response) => {
     blog.likes = 0
   }
 
-  await blog.save()
-  return response.status(201).json(blog)
+  const { token } = request
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    blog.user = user._id
+
+    const saveBlog = await blog.save()
+    user.blogs = user.blogs.concat(saveBlog._id)
+    await user.save()
+    return response.status(201).json(saveBlog)
+  } catch (exception) {
+    next(exception)
+  }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+blogsRouter.delete('/:id', async (request, response, next) => {
+  try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const blog = await Blog.findById(request.params.id)
 
-  if (!blog) {
-    return response.status(404).json({ error: 'blog not found' })
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+    if (decodedToken.id.toString() !== blog.user.toString()) {
+      return response.status(401).json({ error: 'wrong user' })
+    }
+
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  } catch (exception) {
+    next(exception)
   }
-
-  await Blog.findByIdAndRemove(request.params.id)
-  return response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
